[refactor] 소설 페이징 전략 패키지 구조, 이름 개선 - #19
