--- novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/cursor/NovelCursor.java
@@ -1,4 +1,0 @@
-package com.iucyh.novelservice.novel.repository.query.cursor;
-
-public interface NovelCursor {
-}
