+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/NovelCursorPagingStrategy.java
@@ -1,12 +1,12 @@
-package com.iucyh.novelservice.novel.repository.query.pagingquery;
+package com.iucyh.novelservice.novel.repository.query.paging;
 
 import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
-import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
+import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
-public interface NovelPagingQuery {
+public interface NovelCursorPagingStrategy {
 
     JPAQuery<? extends NovelQueryDto> createQuery(JPAQueryFactory queryFactory, NovelCursor cursor);
     NovelCursor createCursor(NovelQueryDto lastResult);
