+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/NovelLastUpdateCursorPagingStrategy.java
@@ -1,12 +1,12 @@
-package com.iucyh.novelservice.novel.repository.query.pagingquery.strategy;
+package com.iucyh.novelservice.novel.repository.query.paging;
 
 import com.iucyh.novelservice.novel.domain.Novel;
 import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
 import com.iucyh.novelservice.novel.repository.query.dto.QNovelSimpleQueryDto;
-import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
-import com.iucyh.novelservice.novel.repository.query.cursor.NovelViewCountCursor;
+import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
+import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelLastUpdateCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
-import com.iucyh.novelservice.novel.repository.query.pagingquery.NovelPagingQueryBaseTemplate;
+import com.iucyh.novelservice.novel.repository.query.paging.template.AbstractNovelCursorPagingStrategy;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Component;
 import static com.iucyh.novelservice.novel.domain.QNovel.novel;
 
 @Component
-public class NovelViewCountPagingQuery extends NovelPagingQueryBaseTemplate {
+public class NovelLastUpdateCursorPagingStrategy extends AbstractNovelCursorPagingStrategy {
 
     @Override
     protected JPAQuery<? extends NovelQueryDto> createBaseQuery(JPAQueryFactory queryFactory) {
@@ -26,31 +26,31 @@ public class NovelViewCountPagingQuery extends NovelPagingQueryBaseTemplate {
     }
 
     @Override
-    protected OrderSpecifier<?>[] createOrderSpecifiers() {
+    protected OrderSpecifier<?>[] applyOrder() {
         return new OrderSpecifier[] {
-                novel.totalViewCount.desc(),
+                novel.lastEpisodeAt.desc(),
                 novel.id.desc()
         };
     }
 
     @Override
-    protected BooleanExpression createCursorPredicate(NovelCursor cursor) {
-        NovelViewCountCursor novelViewCountCursor = (NovelViewCountCursor) cursor;
-        return novel.totalViewCount.lt(novelViewCountCursor.lastTotalViewCount())
+    protected BooleanExpression applyCursor(NovelCursor cursor) {
+        NovelLastUpdateCursor lastUpdateCursor = (NovelLastUpdateCursor) cursor;
+        return novel.lastEpisodeAt.lt(lastUpdateCursor.lastUpdateDate())
                 .or(
-                        novel.totalViewCount.eq(novelViewCountCursor.lastTotalViewCount())
-                                .and(novel.id.lt(novelViewCountCursor.lastNovelId()))
+                        novel.lastEpisodeAt.eq(lastUpdateCursor.lastUpdateDate())
+                                .and(novel.id.lt(lastUpdateCursor.lastNovelId()))
                 );
     }
 
     @Override
     public NovelCursor createCursor(NovelQueryDto lastResult) {
         Novel lastNovel = lastResult.getNovel();
-        return new NovelViewCountCursor(lastNovel.getId(), lastNovel.getTotalViewCount());
+        return new NovelLastUpdateCursor(lastNovel.getId(), lastNovel.getLastEpisodeAt());
     }
 
     @Override
     public NovelSortType getSupportedSortType() {
-        return NovelSortType.VIEW_COUNT;
+        return NovelSortType.LAST_UPDATE;
     }
 }
