+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/NovelViewCountCursorPagingStrategy.java
@@ -1,12 +1,12 @@
-package com.iucyh.novelservice.novel.repository.query.pagingquery.strategy;
+package com.iucyh.novelservice.novel.repository.query.paging;
 
 import com.iucyh.novelservice.novel.domain.Novel;
 import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
 import com.iucyh.novelservice.novel.repository.query.dto.QNovelSimpleQueryDto;
-import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
-import com.iucyh.novelservice.novel.repository.query.cursor.NovelLikeCountCursor;
+import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
+import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelViewCountCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
-import com.iucyh.novelservice.novel.repository.query.pagingquery.NovelPagingQueryBaseTemplate;
+import com.iucyh.novelservice.novel.repository.query.paging.template.AbstractNovelCursorPagingStrategy;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Component;
 import static com.iucyh.novelservice.novel.domain.QNovel.novel;
 
 @Component
-public class NovelLikeCountPagingQuery extends NovelPagingQueryBaseTemplate {
+public class NovelViewCountCursorPagingStrategy extends AbstractNovelCursorPagingStrategy {
 
     @Override
     protected JPAQuery<? extends NovelQueryDto> createBaseQuery(JPAQueryFactory queryFactory) {
@@ -26,31 +26,31 @@ public class NovelLikeCountPagingQuery extends NovelPagingQueryBaseTemplate {
     }
 
     @Override
-    protected OrderSpecifier<?>[] createOrderSpecifiers() {
+    protected OrderSpecifier<?>[] applyOrder() {
         return new OrderSpecifier[] {
-                novel.likeCount.desc(),
+                novel.totalViewCount.desc(),
                 novel.id.desc()
         };
     }
 
     @Override
-    protected BooleanExpression createCursorPredicate(NovelCursor cursor) {
-        NovelLikeCountCursor likeCountCursor = (NovelLikeCountCursor) cursor;
-        return novel.likeCount.lt(likeCountCursor.lastLikeCount())
+    protected BooleanExpression applyCursor(NovelCursor cursor) {
+        NovelViewCountCursor novelViewCountCursor = (NovelViewCountCursor) cursor;
+        return novel.totalViewCount.lt(novelViewCountCursor.lastTotalViewCount())
                 .or(
-                        novel.likeCount.eq(likeCountCursor.lastLikeCount())
-                                .and(novel.id.lt(likeCountCursor.lastNovelId()))
+                        novel.totalViewCount.eq(novelViewCountCursor.lastTotalViewCount())
+                                .and(novel.id.lt(novelViewCountCursor.lastNovelId()))
                 );
     }
 
     @Override
     public NovelCursor createCursor(NovelQueryDto lastResult) {
         Novel lastNovel = lastResult.getNovel();
-        return new NovelLikeCountCursor(lastNovel.getId(), lastNovel.getLikeCount());
+        return new NovelViewCountCursor(lastNovel.getId(), lastNovel.getTotalViewCount());
     }
 
     @Override
     public NovelSortType getSupportedSortType() {
-        return NovelSortType.LIKE_COUNT;
+        return NovelSortType.VIEW_COUNT;
     }
 }
