+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/cursor/NovelCursor.java
@@ -1,0 +1,4 @@
+package com.iucyh.novelservice.novel.repository.query.paging.cursor;
+
+public interface NovelCursor {
+}
