+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/cursor/NovelPopularCursor.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.novel.repository.query.cursor;
+package com.iucyh.novelservice.novel.repository.query.paging.cursor;
 
 public record NovelPopularCursor(
